+++ realm/realm-library/src/objectServer/java/io/realm/SyncUser.java
@@ -172,9 +172,13 @@ public class SyncUser {
      * @throws IllegalArgumentException if the URL is malformed.
      */
     public static SyncUser login(final SyncCredentials credentials, final String authenticationUrl) throws ObjectServerError {
-        final URL authUrl;
+        URL authUrl;
         try {
             authUrl = new URL(authenticationUrl);
+            // If no path segment is provided append `/auth` which is the standard location.
+            if (authUrl.getPath().equals("")) {
+                authUrl = new URL(authUrl.toString() + "/auth");
+            }
         } catch (MalformedURLException e) {
             throw new IllegalArgumentException("Invalid URL " + authenticationUrl + ".", e);
         }