+++ CHANGELOG.md
@@ -6,8 +6,6 @@
 
 ### Enhancements
 
-* Now `targetSdkVersion` is 25.
-* Now using Gradle 3.4.1
 * The real `RealmMigrationNeededException` is now thrown instead of `IllegalArgumentException` if no migration is provided for a Realm that requires it.
 * Partial implementation of `LinkingObjects`.  There is documentation in `io.realm.annotations.LinkingObjects`.  Internal docs are in `io.realm.processor.Backlink`.
   * Queries on linking objects do not work.  Queries like `were(...).equalTo("field.linkingObjects.id", 7).findAll()` are not yet supported.
@@ -15,8 +13,26 @@
   * Migration for linking objects is not yet supported.
   * Backlink verification is incomplete.  Evil code can cause native crashes.
 * [ObjectServer] In case of a Client Reset, information about the location of the backed up Realm file is now reported through the `ErrorHandler` interface (#4080).
+* [ObjectServer] Authentication URLs now automatically append `/auth` if no other path segment is set (#4370).
 * The listener on `RealmObject` will only be triggered if the object changes (#3894).
 * Added `RealmObjectChangeListener` to get detailed information about `RealmObject` changes.
+
+### Bug Fixes
+
+### Deprecated
+
+### Internal
+
+* Using the Object Store's Session and SyncManager.
+
+
+## 3.0.1 (YYYY-MM-DD)
+
+### Enhancements
+
+* Now using Gradle 3.4.1
+* Now `targetSdkVersion` is 25.
+* Listeners on `RealmList` and `RealmResults` will be triggered immediately when the transaction is committed on the same thread (#4245).
 
 ### Bug Fixes
 
@@ -24,14 +40,15 @@
 * Fixed a potential cause for Realm file corruptions (never reported).
 * Add `@Override` annotation to proxy class accessors and stop using raw type in proxy classes in order to remove warnings from javac (#4329).
 * `findFirstAsync()` now returns an invalid object if there is no object matches the query condition instead of running the query repeatedly until it can find one (#4352).
+* [ObjectServer] Changing the log level after starting a session now works correctly (#4337).
 
 ### Deprecated
 
 ### Internal
 
-* Using the Object Store's Session and SyncManager.
 * Upgraded to Realm Sync 1.3.2.
 * Upgraded to Realm Core 2.4.0.
+
 
 ## 3.0.0 (2017-02-28)
 
@@ -137,7 +154,7 @@
 * Updated to Realm Sync v1.0.0.
 * Added a Realm backup when receiving a Sync client reset message from the server.
 
-## 2.2.2 (2016-01-16)
+## 2.2.2 (2017-01-16)
 
 ### Object Server API Changes (In Beta)